[Refactor] Merge Title and TitleBuilder
It's unclear why we need this to be two separate classes
